backend/requirements.txt
Flask
flask-cors
PyMuPDF
python-docx
spacy
scikit-learn
pandas
numpy
nltk
lime
Werkzeug
